[#264] cookie 관련 로직을 http mod로 이동
