+++ rupring/src/core/mod.rs
@@ -2,7 +2,6 @@ mod banner;
 pub mod boot;
 pub(crate) mod bootings;
 mod compression;
-pub(crate) mod cookie;
 mod error_handler;
 mod graceful;
 mod parse;
@@ -27,6 +26,7 @@ use crate::application_properties::ApplicationProperties;
 use crate::application_properties::CompressionAlgorithm;
 use crate::di;
 use crate::header;
+use crate::http::cookie;
 use crate::http::multipart;
 use crate::request::Metadata;
 pub(crate) mod route;
