+++ rupring/src/http/mod.rs
@@ -1,4 +1,5 @@
 pub mod cache;
+pub mod cookie;
 pub mod header;
 pub mod meme;
 pub mod multipart;
